Exctracted  the ThermalManager into separated module
